Refactor Context and visitors to better handle the various paths in preparation of proper partial template specialization support
