+++ lib/visitors/Var.ts
@@ -1,0 +1,3 @@
+import { Context } from "../Context.ts";
+
+export const visitVar = (context: Context) => {};
